Move date related method to a separate module
Move date related method to a separate module similary as who it is
done in `url_filters.rb`, as `filters.rb` is getting to long and
Rubocop is not happy with it.
